+++ packages/core/test/golden/array-root.plain.zig
@@ -1,6 +1,6 @@
 pub const Users = []const User;
 
 pub const User = struct {
-    id: u64,
+    id: u8,
     name: []const u8,
 };
