+++ packages/core/test/golden/flat.plain.zig
@@ -1,5 +1,5 @@
 pub const User = struct {
-    id: u64,
+    id: u8,
     name: []const u8,
     active: bool,
 };
