+++ packages/core/test/golden/map.plain.zig
@@ -3,5 +3,5 @@ const std = @import("std");
 pub const Cfg = std.StringHashMap(CfgValue);
 
 pub const CfgValue = struct {
-    v: u64,
+    v: u8,
 };
