+++ packages/core/test/golden/nested.plain.zig
@@ -1,5 +1,5 @@
 pub const User = struct {
-    id: u64,
+    id: u8,
     profile: Profile,
 };
