+++ packages/core/test/golden/optional.plain.zig
@@ -1,6 +1,6 @@
 const std = @import("std");
 
 pub const User = struct {
-    id: u64,
+    id: u8,
     email: ?std.json.Value = null,
 };
