+++ packages/serde-zig/test/golden/camel-keys.serde.zig
@@ -1,7 +1,7 @@
 const serde = @import("serde");
 
 pub const User = struct {
-    user_id: u64,
+    user_id: u8,
     first_name: []const u8,
     last_seen: []const u8,
