+++ packages/serde-zig/test/golden/kebab-config.serde.zig
@@ -1,8 +1,8 @@
 const serde = @import("serde");
 
 pub const Cfg = struct {
-    server_port: u64,
-    max_connections: u64,
+    server_port: u16,
+    max_connections: u8,
     log_level: []const u8,
 
     pub const serde = .{
